Fix bug with routing to city page
