+++ src/ErrorMessage.css
@@ -4,11 +4,12 @@
     flex-direction: column; 
     align-items: center; 
     text-align: center;
+    height: 100vh;
 }
 
 .error-img {
-    height: 40%; 
-    width: 40%; 
+    height: 500px; 
+    width: 500px; 
     margin-top: 0; 
     &:hover {
         transform: none;
