+++ src/FilterBar.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import React, { useState } from 'react';
 
 export default function FilterBar({onApplyFilters, resetFilters}) {
 
@@ -7,8 +7,9 @@ export default function FilterBar({onApplyFilters, resetFilters}) {
     const [isSmoker, setIsSmoker] = useState('');
     const [ageMax, setAgeMax] = useState('');
     const [budgetRange, setBudgetRange] = useState('');
-
+  
     const handleApplyFilterClick = () => {
+   
         onApplyFilters({
             gender,
             hasPets,
