+++ src/NotFound.css
@@ -1,7 +1,6 @@
 .error {
-    height: 40%; 
-    width: 40%; 
-    margin-top: 0; 
+    height: 400px; 
+    width: 450px; 
     &:hover {
         transform: none;   
     }
@@ -9,6 +8,7 @@
 
 .error-container {
     background-color: #eec0c8;
+    height: 100vh; 
 }
 
 .error-header {
@@ -16,3 +16,12 @@
     margin-bottom: 0; 
 }
 
+.home-button {
+    margin-top: 4%; 
+    border-radius: 15px;
+    border: none; 
+    width: 250px;
+    height: 30px; 
+    font-size: larger;
+    cursor: pointer; 
+}
