+++ src/NotFound.js
@@ -8,7 +8,7 @@ export default function NotFound(){
             <h2 className='error-header'>404 PAGE NOT FOUND</h2>
             <img className='error' src={Error} />
             <div>
-              <Link to={'/'}><button>Back to Homepage</button></Link>  
+              <Link to={'/'}><button className='home-button'>Back to Homepage</button></Link>  
             </div>
             
         </div>
