+++ src/RoommateDetail.js
@@ -14,13 +14,13 @@ function RoommateDetails ( {allData} ) {
     return (
         <div className="roommate-details">
             <h2>{details.name}, {details.age}</h2>
-            <div classname="detail"><strong>Smoker:</strong> {details.isSmoker ? 'Yes' : 'No'} </div>
-            <div classname="detail"><strong>Pets:</strong> {details.hasPets ? 'Yes' : 'No'} </div>
-            <div classname="detail"><strong>Max Budget: $</strong>{details.maxBudget} </div>
-            <div classname="detail"> <strong>Gender:</strong> {details.gender}</div>
-            <div classname="detail"> <strong>About Me:</strong> {details.bio}</div>
-            <div classname="detail"> <strong>Additional Notes:</strong> {details.important}</div>
-            <div classname="profile-image"> 
+            <div className="detail"><strong>Smoker:</strong> {details.isSmoker ? 'Yes' : 'No'} </div>
+            <div className="detail"><strong>Pets:</strong> {details.hasPets ? 'Yes' : 'No'} </div>
+            <div className="detail"><strong>Max Budget: $</strong>{details.maxBudget} </div>
+            <div className="detail"> <strong>Gender:</strong> {details.gender}</div>
+            <div className="detail"> <strong>About Me:</strong> {details.bio}</div>
+            <div className="detail"> <strong>Additional Notes:</strong> {details.important}</div>
+            <div className="profile-image"> 
                 <img src = {details.image} alt={`Profile pic for ${details.name}`}/>
             </div>
